Move PathPredicate from pktcls to pathmgr. (#1286)
This will prevent the BR from importing `gopacket` via ctrl->sigmgmt->sigcmn->snet->pathmgr->pktcls->gopacket, and therefore reverse the resident memory usage jump from ~20M to ~40M, which is causing CI to fail due to running out of memory.
