+++ go/lib/pathmgr/pathmgr.go
@@ -46,7 +46,6 @@ import (
 	"github.com/netsec-ethz/scion/go/lib/addr"
 	"github.com/netsec-ethz/scion/go/lib/common"
 	"github.com/netsec-ethz/scion/go/lib/ctrl/path_mgmt"
-	"github.com/netsec-ethz/scion/go/lib/pktcls"
 	"github.com/netsec-ethz/scion/go/lib/sciond"
 )
 
@@ -159,7 +158,7 @@ func (r *PR) Register(src, dst *addr.ISD_AS) (*SyncPaths, error) {
 //
 // RegisterFilter also adds pair src-dst to the list of tracked paths (if it
 // wasn't already tracked).
-func (r *PR) RegisterFilter(src, dst *addr.ISD_AS, filter *pktcls.PathPredicate) (*SyncPaths, error) {
+func (r *PR) RegisterFilter(src, dst *addr.ISD_AS, filter *PathPredicate) (*SyncPaths, error) {
 	r.Lock()
 	defer r.Unlock()
 
@@ -213,7 +212,7 @@ func (r *PR) register(src, dst *addr.ISD_AS) (*SyncPaths, error) {
 }
 
 // UnregisterFilter deletes a previously registered filter.
-func (r *PR) UnregisterFilter(src, dst *addr.ISD_AS, filter *pktcls.PathPredicate) error {
+func (r *PR) UnregisterFilter(src, dst *addr.ISD_AS, filter *PathPredicate) error {
 	// TODO(scrye): implement this
 	return common.NewCError("Not implemented")
 }
