+++ go/lib/pathmgr/pathmgr_test.go
@@ -22,7 +22,6 @@ import (
 	"time"
 
 	log "github.com/inconshreveable/log15"
-	. "github.com/smartystreets/goconvey/convey"
 
 	"github.com/netsec-ethz/scion/go/lib/addr"
 )
